pyproject.toml
[project]
name = "nitrotype-auto"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.5"
dependencies = [
    "pyyaml>=6.0.2",
    "selenium>=4.34.2",
]
